refactor: 공통 메서드 Helper 클래스 분리
- OrderHelper (application 폴더로 위치 변경)
- OrderQueue (공통 메서드 분리)
