+++ order-service/src/main/java/com/oingmaryho/business/orderservice/application/service/OrderService.java
@@ -2,6 +2,7 @@ package com.oingmaryho.business.orderservice.application.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.oingmaryho.business.common.domain.type.UserRoleType;
+import com.oingmaryho.business.orderservice.application.OrderHelper;
 import com.oingmaryho.business.orderservice.application.dto.mapper.OrderApplicationMapper;
 import com.oingmaryho.business.orderservice.application.dto.request.*;
 import com.oingmaryho.business.orderservice.application.dto.response.HubSearchResponseDto;
